requirements.txt
opencv-python~=4.5.2.52
numpy~=1.20.2
Pillow~=8.2.0
scikit-image~=0.18.1
pytesseract~=0.3.7